scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "dotenv>=0.9.9",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pymongo>=4.15.5",
    "tqdm>=4.67.1",
]
